PersonalAssistantRaga/perosnalassistantraga/pyproject.toml
[project]
name = "perosnalassistantraga"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "python-dotenv>=1.2.1",
]
